Emmett

+++ database/migrations/2017_09_07_221438_create_lotteries_table.php
@@ -16,8 +16,8 @@ class CreateLotteriesTable extends Migration
             $table->string('content');
             $table->integer('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('init_value');
-            $table->integer('current_value');
+            $table->decimal('init_value');
+            $table->decimal('current_value');
             $table->timestamps();
             $table->dateTime('end_date');
         });
