+++ database/migrations/2017_09_08_231448_create_the_world_lotteries_table.php
@@ -16,8 +16,8 @@ class CreateTheWorldLotteriesTable extends Migration
             $table->string('title');
             $table->integer('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('init_value');
-            $table->integer('current_value');
+            $table->decimal('init_value');
+            $table->decimal('current_value');
             $table->timestamps();
             $table->dateTime('end_date');
         });
